Move diffusion_dpo_lora_sd.py script to _experimental/ directory.
